Create errs package
Create package `errs` which exports `New` and `Errorf` functions, both
allowing to wrap more than a single error. The newly created error is
query-able with `errors.Is` and `errors.As`.

This extension is particularly useful when tagging and error with  a
sentinel error value _and_ keeping the underlying, causing error.

Test with:

	go test ./...