Initialise httpe package
Implement core functionality of httpe package: an http.Handler that can
return an error.

	type HandlerE interface {
	       ServeHTTP(http.ResponseWriter, *http.Request) error
	}

It is intended to be used with ErrWriter which handles the error and
turns it into an appropriate http response StatusCode and Body.

	type ErrWriter interface {
	       Write(http.ResponseWriter, error)
	}

`Chain` chains several HandlerEs sequentially and bails on the first
error.

`NewHandler` creates a traditional http.Handler from a `HandlerE` and
`ErrWriter`.

The httpe package also provides HandlerFuncE and ErrWriterFunc types as
adapters to the interfaces, similar to http.HandlerFunc for http.Handler
interface.

Introduce new `mock` package exposing mock.ResponseWriter which
implements http.ResponseWriter. It captures the bytes and http status
code written and can be preset to respond with an error on Write calls.