Initialise package timex
Create package timex with type Ticker. A Ticker that holds a channel
that delivers 'ticks' of a clock at intervals, similar to time.Ticker
from the standard library.

It differs from the standard library Ticker by closing C when
Stop() is called, which allows for ranging over C:

	for range ticker.C {
  		// do something
	}

In a follow up commit we will add a TickerWithContext() function based
on this new ticker that stops when the context is cancelled.