timeout: Don't kill with SIGKILL on timeout
The original version of `timeout` used a context with a timeout to have
`exec.Cmd` kill the running command. However, that uses `SIGKILL` which
does not give the child process a chance to clean up. In the case where
this was used, `gcloud` had forked a child process. When the parent was
killed with `SIGKILL` it was not able to clean up the child.

Now `SIGTERM` is sent and if the process has not exited within a given
grace period (3s by default), then it is killed with `SIGKILL`. The
signal and grace period can be set with `-s` and `-g` respectively.