timeout: Add shell-driven tests
Add `test.sh` to "blackbox" test the timeout command. It is possible
this would be flaky under heavy load as timeout measurements could be
pushed out past the expected number of seconds.

Add the run of the test script to the `test` target of the Makefile,
adding a `build` dependency too so we know the binary exists.