test: Add some environment test helpers
Add some helpers for setting up the OS environment in tests and
restoring it after the test is done.

A type `test.Environ` is added that can set and unset the OS
environment, keeping a record of what changes it makes. At the end of a
test, the `Restore()` method can be called to return the environment
back to the starting state.

A global variable `test.Env` is provided for simplicity as the OS
environment itself is global, it is not usually necessary to instantiate
the `test.Environ` struct.