ci: Fix pkg.go.dev tag pull
The previous commit added a step to the version bump to `go get` @latest
of the Go module so it will appear in the docs at pkg.go.dev. But it
seemed to get the old version with @latest, so use the tag just set
explicitly and see if that works, of if there is a race condition with
github setting the tag and fetching the latest tag.