httpe: Unify NewHandler/NewSafeHandler with options
Use the options pattern to unify `NewHandler` and `NewSafeHandler` (and
their `Func` siblings) so that the default is now to use `WriteSafeErr`
to write errors to the `http.ResponseWriter` and allow it to be
overridden with an option `WithErrWriter`/`WithErrWriterFunc`.

The make simple use simple, keeping a tighter but extensible API
surface.

Link: https://dave.cheney.net/2014/10/17/functional-options-for-friendly-apis
Link: https://commandcenter.blogspot.com/2014/01/self-referential-functions-and-design.html