httpe: Add New() function for magical chaining
Add a `New()` function that accepts various types of parameters
representing handlers and chain them together calling them one after the
other. Each is treated as a HandlerE so can return errors. If any
handler in the chain does return an error, no further handlers in the
chain are called, and the error is passed to an ErrWriter.

One ErrWriter can be supplied in the argument list (anywhere), and that
will be used as the ErrWriter to handle errors. If none is supplied,
then the default `WriteSafeErr` is used.

The types of handlers accepted are HandlerE and http.Handler (which
includes http.HandlerFunc and HandlerFuncE), and also plain functions
that match the HandlerFunc types. Same for the ErrWriter arg - it can be
a function that matches ErrWriterFunc.

This dynamic typing makes the use when mixing http.Handler and plain
functions much nicer:

    h, err := New(authFunc, corsAllowAll, httpe.Get, handle, writeErr)

where you can mix and match the types without needing to cast to
HandlerFuncE, or adapt http.Handler into HandlerE.

Add tests and an example for the new code.