Revert "errs: Add Ignore function"
The errs.Ignore function did not end up working the way we wanted it to,
as lint could no longer tell that `Close` was being called on
`Request.Body` since the `Close` function was passed by value to be
executed anonymously.

Since the main purpose was to have code that documented explicitly that
we were ignoring errors and not have lint warnings, and that did not
work, there is no purpose to this code any more.

This reverts commit 1ed685a7ff1ca894d766b17ac65f38e0ecb1ca2e.