github: Unpin setup-go version, just use "v2"
Change the version of the setup-go action to v2 from v2.1.2. Later
versions fix problems and we want those fixes automatically. In this
case, GitHub has disables the `set-env` and `add-path` commands which
v2.1.2 used. We need a newer version. Rather than bumping to a new point
release, just use "v2". (my personal preference is for reproducible
builds, so we used a specific version, but that is not really necessary
as we can reproduce a build manually).